Revert "move avatar output (i.e., configurable machine) behind log flag"
(this is a temporary revert to remove the log_mask changes from int ->
long)

This reverts commit 4ca68de18eea0f67c97e3731aff570715339b18f.

+++ panda/plugins/taint2/shad.h
@@ -57,7 +57,7 @@ extern uint32_t max_taintset_card;
 #define tassert(cond) assert((cond))
 #define taint_log(...) qemu_log_mask(CPU_LOG_TAINT_OPS, ## __VA_ARGS__);
 #define taint_log_labels(shad, src, size) \
-    extern long qemu_loglevel; \
+    extern int qemu_loglevel; \
     if (qemu_loglevel & CPU_LOG_TAINT_OPS) { \
         bool tainted = false; \
         for (int __i = 0; __i < size; __i++) { \
